+++ .gitignore
@@ -14,6 +14,7 @@ dist/
 downloads/
 eggs/
 .eggs/
+database.db
 lib/
 lib64/
 parts/
