+++ sniffer.py
@@ -3,9 +3,27 @@ import threading
 import time
 import select
 import uuid
-from database import cur
 import logging
 import sys
+
+
+import os
+import sqlite3
+
+DB_PATH = 'database.db'
+db_exists = os.path.exists(DB_PATH)
+db = sqlite3.connect(DB_PATH, check_same_thread=False, isolation_level=None)
+cur = db.cursor()
+
+if not db_exists:
+    cur.execute('''CREATE TABLE data (
+                session_id TEXT,
+                timestamp UNSIGNED BIGINT,
+                direction INT NOT NULL,
+                data BLOB NOT NULL,
+                PRIMARY KEY (session_id, timestamp)
+                )''')
+
 
 
 logger = logging.getLogger(__name__)
@@ -35,6 +53,7 @@ class ChatServer(threading.Thread):
     def save_data(self, direction: int, data):
         cur.execute('INSERT INTO data (session_id, timestamp, direction, data) VALUES (?, ?, ?, ?)',
                     (self.id, time.time_ns()//1000, direction, data))
+        db.commit()
 
     def run(self):
         while True:
